+++ core/render2d/render_backend.h
@@ -56,14 +56,14 @@ namespace indigo
         // ---- Path operations ----
         virtual void beginPath() = 0;
         virtual void closePath() = 0;
-        virtual void moveTo(float x, float y) = 0;
-        virtual void lineTo(float x, float y) = 0;
-        virtual void relMoveTo(float dx, float dy) = 0;
-        virtual void relLineTo(float dx, float dy) = 0;
-        virtual void curveTo(float x1, float y1, float x2, float y2, float x3, float y3) = 0;
-        virtual void arc(float cx, float cy, float r, float a0, float a1) = 0;
-        virtual void arcNegative(float cx, float cy, float r, float a0, float a1) = 0;
-        virtual void rect(float x, float y, float w, float h) = 0;
+        virtual void moveTo(double x, double y) = 0;
+        virtual void lineTo(double x, double y) = 0;
+        virtual void relMoveTo(double dx, double dy) = 0;
+        virtual void relLineTo(double dx, double dy) = 0;
+        virtual void curveTo(double x1, double y1, double x2, double y2, double x3, double y3) = 0;
+        virtual void arc(double cx, double cy, double r, double a0, double a1) = 0;
+        virtual void arcNegative(double cx, double cy, double r, double a0, double a1) = 0;
+        virtual void rect(double x, double y, double w, double h) = 0;
 
         // ---- Drawing operations ----
         virtual void fill() = 0;
@@ -71,9 +71,9 @@ namespace indigo
         virtual void paint() = 0;
 
         // ---- Style ----
-        virtual void setSourceRGB(float r, float g, float b) = 0;
-        virtual void setSourceRGBA(float r, float g, float b, float a) = 0;
-        virtual void setLineWidth(float w) = 0;
+        virtual void setSourceRGB(double r, double g, double b) = 0;
+        virtual void setSourceRGBA(double r, double g, double b, double a) = 0;
+        virtual void setLineWidth(double w) = 0;
         virtual void setLineJoin(int join) = 0; // 0=miter, 1=round, 2=bevel
         virtual void setDash(const double* pattern, int count, double offset) = 0;
         virtual void setOperator(int op) = 0; // CAIRO_OPERATOR_SOURCE = 1
@@ -82,9 +82,9 @@ namespace indigo
         // ---- Transform ----
         virtual void save() = 0;
         virtual void restore() = 0;
-        virtual void translate(float dx, float dy) = 0;
-        virtual void scale(float sx, float sy) = 0;
-        virtual void rotate(float angle) = 0;
+        virtual void translate(double dx, double dy) = 0;
+        virtual void scale(double sx, double sy) = 0;
+        virtual void rotate(double angle) = 0;
 
         // Matrix operations: get/set current transform matrix as 6 doubles [xx, yx, xy, yy, x0, y0]
         virtual void getMatrix(double m[6]) = 0;
@@ -101,8 +101,8 @@ namespace indigo
 
         // ---- Text ----
         virtual void selectFontFace(const char* family, bool italic, bool bold) = 0;
-        virtual void setFontSize(float size) = 0;
-        virtual void textExtents(const char* text, float& width, float& height, float& x_bearing, float& y_bearing) = 0;
+        virtual void setFontSize(double size) = 0;
+        virtual void textExtents(const char* text, double& width, double& height, double& x_bearing, double& y_bearing) = 0;
         virtual void fontExtents(double& height) = 0;
         virtual void showText(const char* text) = 0;
         virtual void textPath(const char* text) = 0;
@@ -114,12 +114,12 @@ namespace indigo
         virtual void applyFontOptions() = 0;
 
         // ---- Gradient ----
-        virtual void setLinearGradient(float x0, float y0, float x1, float y1, float r1, float g1, float b1, float r2, float g2, float b2) = 0;
+        virtual void setLinearGradient(double x0, double y0, double x1, double y1, double r1, double g1, double b1, double r2, double g2, double b2) = 0;
         virtual void clearPattern() = 0;
 
         // ---- Image operations ----
         // Draw a PNG image from raw data into the given bounding box
-        virtual void drawPngImage(const void* data, int dataLen, float x, float y, float w, float h) = 0;
+        virtual void drawPngImage(const void* data, int dataLen, double x, double y, double w, double h) = 0;
         // Write surface to PNG stream
         virtual void writeSurfaceToPng(void* output) = 0;
 
@@ -127,7 +127,7 @@ namespace indigo
         virtual bool isPathEmpty() = 0;
 
         // ---- Surface source ----
-        virtual void setSourceSurface(float x, float y) = 0;
+        virtual void setSourceSurface(double x, double y) = 0;
     };
 
 } // namespace indigo
